refactor: fix typo in filename
